+++ blacklist/blacklist/Scenes/Home/HomeViewController.swift
@@ -15,10 +15,6 @@ class HomeViewController: UIViewController, ViewControllerProtocol {
     var presenter: HomePresenter!
     var router: HomeRouter?
 
-    struct Constants {
-        static let upcomingTableViewCell = "UpcomingTableViewCell"
-    }
-
     fileprivate var tableViewTopConstraintConstant: CGFloat = 0.0
 
     // IBOutlets
@@ -84,8 +80,8 @@ class HomeViewController: UIViewController, ViewControllerProtocol {
 extension HomeViewController: UITableViewDataSource, UITableViewDelegate {
     func configureTableView() {
         // Let's configure the tableView
-        let upcomingTableViewCellNib = UINib(nibName: Constants.upcomingTableViewCell, bundle: nil)
-        tableView.register(upcomingTableViewCellNib, forCellReuseIdentifier: Constants.upcomingTableViewCell)
+        let upcomingTableViewCellNib = UINib(nibName: UpcomingTableViewCell.identifier, bundle: nil)
+        tableView.register(upcomingTableViewCellNib, forCellReuseIdentifier: UpcomingTableViewCell.identifier)
 
         tableView.estimatedRowHeight = 68.0
         tableView.rowHeight = UITableViewAutomaticDimension
@@ -107,7 +103,7 @@ extension HomeViewController: UITableViewDataSource, UITableViewDelegate {
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(
-            withIdentifier: Constants.upcomingTableViewCell,
+            withIdentifier: UpcomingTableViewCell.identifier,
             for: indexPath
         ) as! UpcomingTableViewCell
